requirements.txt
wagtail==7.2.1
Django==6.0.1
pyclean==3.4.0
django-countries==8.2.0
django-allauth==65.13.1
django-widget-tweaks==1.5.0
wagtail-modeladmin==2.2.0
django-extensions==4.1
lorem-text==3.0
django-silk==5.4.3
djangorestframework_simplejwt==5.5.1